Moving some tests
